pyproject.toml
[project]
name = "notifications"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "enums>=0.0.2",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "tortoise-orm>=1.1.4",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.2",
]
